cflat_runtime: match shift signedness in objectFrame sign-compare opcodes
For the xor-based less-than opcodes (0x2E/0x30), the target uses an
arithmetic (value>>1) and a logical final >>31. Reorder the casts so
value>>1 is signed and the result is shifted as unsigned.
objectFrame 66.28% -> 66.37%.

+++ src/cflat_runtime.cpp
@@ -1956,7 +1956,7 @@ frameLoop:
 			unsigned int* sp = object->m_sp;
 			--object->m_sp;
 			const u32 value = sp[-1] ^ sp[-2];
-			sp[-2] = (static_cast<int>((value >> 1) - (value & sp[-1])) >> 31);
+			sp[-2] = static_cast<u32>((static_cast<int>(value) >> 1) - static_cast<int>(value & sp[-1])) >> 31;
 			break;
 		}
 		case 0x2F: {
@@ -1970,7 +1970,7 @@ frameLoop:
 			unsigned int* sp = object->m_sp;
 			--object->m_sp;
 			const u32 value = sp[-2] ^ sp[-1];
-			sp[-2] = (static_cast<int>((value >> 1) - (value & sp[-2])) >> 31);
+			sp[-2] = static_cast<u32>((static_cast<int>(value) >> 1) - static_cast<int>(value & sp[-2])) >> 31;
 			break;
 		}
 		case 0x31: {
